# docs/swagger.yaml
definitions:
  errorCode.ErrorCode:
    enum:
    - 0
    - 1001
    - 1002
    - 1003
    - 1004
    format: int32
    type: integer
    x-enum-varnames:
    - UnknownErrCode
    - UnauthorizedErrCode
    - InternalErrCode
    - NotFoundErrCode
    - NoAccessErrCode
...
  service.ErrorCodeResponse:
    properties:
      requestUUID:
        type: string
      subcode:
        $ref: '#/definitions/errorCode.ErrorCode'
    type: object
  service.Friend:
    properties:
      id:
        type: integer
      name:
        type: string
      photo_url:
        type: string
      username:
        type: string
    type: object
...
  service.IncomingFriendsRequestsCountResponse:
    properties:
      count:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend service for wishlists app
  title: Wishlists backend API
  version: "1.0"
paths:
...
  /api/user/friend/request:
    post:
      parameters:
      - description: Friend username
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/service.ErrorCodeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/service.ErrorCodeResponse'
      security:
      - ApiKeyAuth: []
      summary: creates a friend request to another user
      tags:
      - Friends requests
  /api/user/friend/request/accept:
    post:
      parameters:
      - description: Friend ID
        in: query
        name: friend_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/service.ErrorCodeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/service.ErrorCodeResponse'
      security:
      - ApiKeyAuth: []
      summary: accepts an incoming friend request
      tags:
      - Friends requests
  /api/user/friend/request/deny:
    post:
      parameters:
      - description: Friend ID
        in: query
        name: friend_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/service.ErrorCodeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/service.ErrorCodeResponse'
      security:
      - ApiKeyAuth: []
      summary: accepts an incoming friend request
      tags:
      - Friends requests
...
  /api/user/friends/requests/incoming:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.Friend'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/service.ErrorCodeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/service.ErrorCodeResponse'
      security:
      - ApiKeyAuth: []
      summary: returns user's incoming friends requests
      tags:
      - Friends requests
  /api/user/friends/requests/incoming/count:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.IncomingFriendsRequestsCountResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/service.ErrorCodeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/service.ErrorCodeResponse'
      security:
      - ApiKeyAuth: []
      summary: returns user's incoming friends requests count
      tags:
      - Friends requests
  /api/user/friends/requests/outcoming:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.Friend'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/service.ErrorCodeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/service.ErrorCodeResponse'
      security:
      - ApiKeyAuth: []
      summary: returns user's incoming friends requests
      tags:
      - Friends requests
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"